# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SpotStatus = (typeof SpotStatus)[keyof typeof SpotStatus];

export const SpotStatus = {
  available: "available",
  occupied: "occupied",
  reserved: "reserved",
  maintenance: "maintenance",
} as const;

export type SpotType = (typeof SpotType)[keyof typeof SpotType];

export const SpotType = {
  standard: "standard",
  compact: "compact",
  accessible: "accessible",
  ev: "ev",
  motorcycle: "motorcycle",
} as const;

export interface Spot {
  id: string;
  code: string;
  zone: string;
  level: number;
  type: SpotType;
  status: SpotStatus;
  hourlyRate: number;
  createdAt: string;
}

export interface SpotInput {
  code: string;
  zone: string;
  level: number;
  type: SpotType;
  hourlyRate: number;
  status?: SpotStatus;
}

export interface SpotUpdate {
  code?: string;
  zone?: string;
  level?: number;
  type?: SpotType;
  hourlyRate?: number;
  status?: SpotStatus;
}

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  upcoming: "upcoming",
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Reservation {
  id: string;
  spotId: string;
  spotCode: string;
  vehicleId: string;
  vehiclePlate: string;
  startTime: string;
  endTime: string;
  status: ReservationStatus;
  totalCost: number;
  createdAt: string;
}

export interface ReservationInput {
  spotId: string;
  vehicleId: string;
  startTime: string;
  endTime: string;
}

export interface ReservationUpdate {
  status?: ReservationStatus;
  startTime?: string;
  endTime?: string;
}

export interface Vehicle {
  id: string;
  plate: string;
  make: string;
  model: string;
  color: string;
  type: SpotType;
  ownerName?: string;
  createdAt: string;
}

export interface VehicleInput {
  plate: string;
  make: string;
  model: string;
  color: string;
  type: SpotType;
  ownerName?: string;
}

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  paid: "paid",
  pending: "pending",
  refunded: "refunded",
} as const;

export interface Transaction {
  id: string;
  reservationId: string;
  spotCode: string;
  vehiclePlate: string;
  durationMinutes: number;
  amount: number;
  paidAt: string;
  status: TransactionStatus;
}

export interface DashboardSummary {
  totalSpots: number;
  availableSpots: number;
  occupiedSpots: number;
  reservedSpots: number;
  maintenanceSpots: number;
  utilizationRate: number;
  revenueToday: number;
  revenueWeek: number;
  activeReservations: number;
  totalVehicles: number;
}

export type ActivityItemKind =
  (typeof ActivityItemKind)[keyof typeof ActivityItemKind];

export const ActivityItemKind = {
  reservation_created: "reservation_created",
  reservation_completed: "reservation_completed",
  spot_occupied: "spot_occupied",
  spot_freed: "spot_freed",
  vehicle_added: "vehicle_added",
} as const;

export interface ActivityItem {
  id: string;
  kind: ActivityItemKind;
  message: string;
  spotCode?: string;
  vehiclePlate?: string;
  at: string;
}

export interface ZoneStat {
  zone: string;
  total: number;
  available: number;
  occupied: number;
  reserved: number;
}

export interface RevenuePoint {
  date: string;
  revenue: number;
  sessions: number;
}

export type ListSpotsParams = {
  zone?: string;
  status?: SpotStatus;
};

export type ListReservationsParams = {
  status?: ReservationStatus;
  vehicleId?: string;
};

export type ListTransactionsParams = {
  vehicleId?: string;
  /**
   * @minimum 1
   * @maximum 200
   */
  limit?: number;
};