# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemKind } from "./activityItemKind";

export interface ActivityItem {
  id: string;
  kind: ActivityItemKind;
  message: string;
  spotCode?: string;
  vehiclePlate?: string;
  at: Date;
}