# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalSpots: number;
  availableSpots: number;
  occupiedSpots: number;
  reservedSpots: number;
  maintenanceSpots: number;
  utilizationRate: number;
  revenueToday: number;
  revenueWeek: number;
  activeReservations: number;
  totalVehicles: number;
}