# lib/api-zod/src/generated/types/reservationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationInput {
  spotId: string;
  vehicleId: string;
  startTime: Date;
  endTime: Date;
}