# lib/api-zod/src/generated/types/reservationStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  upcoming: "upcoming",
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;