# lib/api-zod/src/generated/types/revenuePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */

export interface RevenuePoint {
  date: Date;
  revenue: number;
  sessions: number;
}