# lib/api-zod/src/generated/types/spot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */
import type { SpotStatus } from "./spotStatus";
import type { SpotType } from "./spotType";

export interface Spot {
  id: string;
  code: string;
  zone: string;
  level: number;
  type: SpotType;
  status: SpotStatus;
  hourlyRate: number;
  createdAt: Date;
}