# lib/api-zod/src/generated/types/spotInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */
import type { SpotStatus } from "./spotStatus";
import type { SpotType } from "./spotType";

export interface SpotInput {
  code: string;
  zone: string;
  level: number;
  type: SpotType;
  hourlyRate: number;
  status?: SpotStatus;
}