# lib/api-zod/src/generated/types/spotType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */

export type SpotType = (typeof SpotType)[keyof typeof SpotType];

export const SpotType = {
  standard: "standard",
  compact: "compact",
  accessible: "accessible",
  ev: "ev",
  motorcycle: "motorcycle",
} as const;