# lib/api-zod/src/generated/types/vehicleInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking System API
 * OpenAPI spec version: 0.1.0
 */
import type { SpotType } from "./spotType";

export interface VehicleInput {
  plate: string;
  make: string;
  model: string;
  color: string;
  type: SpotType;
  ownerName?: string;
}